refactor: move status logic tests to helpers_test.go
The tests for validateStatusFlags() and normalizeDisplayFlags() were in
status_logic_test.go, but these functions are defined in helpers.go.
Moved the tests to helpers_test.go where they belong and removed the
confusing status_logic_test.go file.
